Package-ified QR-Auth Android Client version.
Awaiting Bintray to load the component, and voilà !
The backend is placed in the client_backend/ section.
Not integrated with Travis.
