--- AndroidApp/qr-authserver/build.gradle
@@ -1,34 +1,0 @@
-apply plugin: 'com.android.library'
-
-android {
-    compileSdkVersion 26
-    buildToolsVersion "26.0.1"
-
-    defaultConfig {
-        minSdkVersion 15
-        targetSdkVersion 26
-        versionCode 1
-        versionName "1.0"
-
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-}
-
-dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
-        exclude group: 'com.android.support', module: 'support-annotations'
-    })
-    compile 'com.android.support:appcompat-v7:26.+'
-    compile 'com.github.nkzawa:socket.io-client:0.3.0'
-    compile 'me.dm7.barcodescanner:zxing:1.9.8'
-    compile 'org.jbundle.util.osgi.wrapped:org.jbundle.util.osgi.wrapped.org.apache.http.client:4.1.2'
-    testCompile 'junit:junit:4.12'
-}
