--- AndroidApp/qr-authserver/src/main/AndroidManifest.xml
@@ -1,10 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-
-    package="fr.jserv.qr_authserver">
-
-    <application android:allowBackup="true" android:label="@string/app_name"
-        android:supportsRtl="true">
-
-    </application>
-
-</manifest>
