+++ client_backend/.idea/modules.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/clientBackend/clientBackend.iml" filepath="$PROJECT_DIR$/clientBackend/clientBackend.iml" />
+      <module fileurl="file://$PROJECT_DIR$/client_backend.iml" filepath="$PROJECT_DIR$/client_backend.iml" />
+    </modules>
+  </component>
+</project>
