+++ client_backend/clientBackend/proguard-rules.pro
